requirements.txt
google-genai
sentence-transformers
faiss-cpu
numpy
fastapi
uvicorn[standard]
pydantic
python-dotenv